script/csv files: updated string algo / regenerated files
Updated the wisben script to align with the correction
given. In its current state the script generated unique
strings which looked like "AAAABCDxxx...xxx", with the
'BCD' being the characters computed from the unique
number generated by the script. These actually need to be
in the form of "BCDAAAAxxx...xxx". This commit corrects the string,
so the second is always the case.

Re ran the benchmarks for each of the files.